Change class Protocol value type from int to unsigned short.

+++ src/protocol.cpp
@@ -12,13 +12,13 @@
 
 #include <ostream>      // std::ostream
 
-Network::Protocol::Protocol(Family t_family, int t_value) :
+Network::Protocol::Protocol(Family t_family, unsigned short t_value) :
     Integer(t_value),
     m_family(t_family)
 {
 }
 
-Network::Protocol& Network::Protocol::operator=(int t_value)
+Network::Protocol& Network::Protocol::operator=(unsigned short t_value)
 {
     m_value = t_value;
     return *this;
